--- inlang/source-code/paraglide/paraglide-unplugin/src/virtual.test.ts
@@ -1,9 +1,0 @@
-// @ts-ignore
-import test from "$virtual/file1.js"
-import { describe, it, expect } from "vitest"
-
-describe("virtual", () => {
-	it("works", () => {
-		expect(test).toBe(true)
-	})
-})
